Code Cleanup (Remove unnecessary property)
I had originally added the `relaxed = true` bit to make the mocking
easier, but testing it now shows it's not necessary. I don't know
whether it wasn't necessary to begin with or I just hadn't thought of
trying it one more time back then, but one way or another, it's not
needed now.